In ApacheConf#include_files, check for abs paths
If the path is absolute, just use what was passed, otherwise build an
absolute path using `@conf_dir`.

Fixes #1013

I've got what seemed to me to be the necessary ingredients for a
successful unit test. I added the mock file and referenced it in
`helper.rb` both in the commands hash and files hash. Using
`binding.pry` I saw that the path was valid when it got sent through the
`mockfile` lamda, but the test is erroring when it gets to
`ApacheConf#read_content` at line 79 where `read_file(to_read[0])` is
called. Apparently at that point it can't find the file.

Could someone with more familiarity with the mocking system tell me what
I'm missing there? :)