Consolidate parsing into object initializer
This class provided several avenues for re-attempting content parsing,
which skirts around the intended design where resources that must be
skipped raise exceptions, which are rescued and handled outside the
class.

To address this, all of the sort of lazy loading and memoization was
removed and all of the content processing now happens when creating a
new instance of the class.

Signed-off-by: Trevor Bramble <tbramble@chef.io>